* commit '53368612e0b01998d4ae5302055928c07786c5d5':
  ays & action fixes
  WIP:pip installation bug fix + pip module support for mac
  bug fix with wget + package handling for unsupportecd brew packages
  homeDir bug fix + env PATH fix
  update sals/Home.md
  docs move
  ays fixes
  bug fix , env not set in mac condition

+++ lib/JumpScale/tools/cuisine/CuisineInstallerDevelop.py
@@ -55,7 +55,12 @@ class CuisineInstallerDevelop:
     @actionrun(action=True)
     def pip(self):
         self.cuisine.installer.base()
-        self.python()
+        
+        if self.cuisine.core.isMac:
+            self.cuisine.core.run("easy_install pip")
+            self.python()
+            return
+        
         C="""
             #important remove olf pkg_resources, will conflict with new pip
             rm -rf /usr/lib/python3/dist-packages/pkg_resources
@@ -68,6 +73,7 @@ class CuisineInstallerDevelop:
         C="cd $tmpDir/;python3 get-pip.py"
         C=self.cuisine.core.args_replace(C)
         self.cuisine.core.run(C)
+
 
     @actionrun(action=True)
     def installJS8Deps(self):
